+++ example/pages/home/index.js
@@ -122,6 +122,10 @@ const menus = [
             {
                 label: 'Popup',
                 to: '/popup'
+            },
+            {
+                label: 'Pull To Refresh',
+                to: '/ptr'
             }
         ]
     },