+++ src/index.js
@@ -31,6 +31,7 @@ import { Picker, CityPicker, PickerGroup } from './components/picker';
 import Badge from './components/badge';
 //non standard
 import { Popup, PopupHeader } from './components/popup';
+import PullToRefresh from './components/ptr';
 
 export default {
     version,
@@ -116,5 +117,6 @@ export default {
 
     //non-standard
     Popup,
-    PopupHeader
+    PopupHeader,
+    PullToRefresh
 };